Tweak: Enemy Spawn, Movement
Changed spawn sequence to one that ramps up in difficulty over time.
Changed player and enemy movement to use rigidbody physics to allow enemies to collide with each other instead of overlapping.
Enemies that collide with player or shield now deal damage respectively before being destroyed.
Added more comments to code.

+++ Assets/Scripts/Projectile.cs
@@ -6,7 +6,7 @@ public class Projectile : MonoBehaviour
     [SerializeField] private float duration = 3f;
 
     [Space]
-    [SerializeField] private int enemyShotDamage = 10;
+    [SerializeField] private float enemyShotDamage = 10;
 
     private PlayerControl player;
     private GameManager gameManager;
